Update Anykernel
Merge all changes till 25th march

+++ out/tools/ak2-core.sh
@@ -27,12 +27,15 @@ contains() { test "${1#*$2}" != "$1" && return 0 || return 1; }
 
 # dump boot and extract ramdisk
 dump_boot() {
+  if [ ! -e "$(echo $block | cut -d\  -f1)" ]; then
+    ui_print " "; ui_print "Invalid partition. Aborting..."; exit 1;
+  fi;
   if [ -f "$bin/nanddump" ]; then
     $bin/nanddump -f /tmp/anykernel/boot.img $block;
   else
     dd if=$block of=/tmp/anykernel/boot.img;
   fi;
-  if [ -f "$bin/unpackelf" ]; then
+  if [ -f "$bin/unpackelf" -a "$($bin/unpackelf -i /tmp/anykernel/boot.img -h -q 2>/dev/null; echo $?)" == 0 ]; then
     $bin/unpackelf -i /tmp/anykernel/boot.img -o $split_img;
     mv -f $split_img/boot.img-ramdisk.cpio.gz $split_img/boot.img-ramdisk.gz;
   else
@@ -92,7 +95,7 @@ write_boot() {
     secondoff=`cat *-secondoff`;
     secondoff="--second_offset $secondoff";
   fi;
-  for i in zImage zImage-dtb Image.gz Image.gz-dtb; do
+  for i in zImage zImage-dtb Image.gz Image.gz-dtb Image Image-dtb Image.bz2 Image.bz2-dtb Image.lzo Image.lzo-dtb Image.lzma Image.lzma-dtb Image.xz Image.xz-dtb Image.lz4 Image.lz4-dtb Image.fit; do
     if [ -f /tmp/anykernel/$i ]; then
       kernel=/tmp/anykernel/$i;
       break;
@@ -102,14 +105,18 @@ write_boot() {
     kernel=`ls *-zImage`;
     kernel=$split_img/$kernel;
   fi;
-  if [ -f /tmp/anykernel/dtb ]; then
-    dtb="--dt /tmp/anykernel/dtb";
-  elif [ -f *-dtb ]; then
+  for i in dtb dt.img; do
+    if [ -f /tmp/anykernel/$i ]; then
+      dtb="--dt /tmp/anykernel/$i";
+      break;
+    fi;
+  done;
+  if [ ! "$dtb" -a -f *-dtb ]; then
     dtb=`ls *-dtb`;
     dtb="--dt $split_img/$dtb";
   fi;
   if [ -f "$bin/mkbootfs" ]; then
-    $bin/mkbootfs /tmp/anykernel/ramdisk | gzip > /tmp/anykernel/ramdisk-new.cpio.gz;
+    $bin/mkbootfs $ramdisk | gzip > /tmp/anykernel/ramdisk-new.cpio.gz;
   else
     cd $ramdisk;
     find . | cpio -H newc -o | gzip > /tmp/anykernel/ramdisk-new.cpio.gz;
@@ -117,8 +124,8 @@ write_boot() {
   if [ $? != 0 ]; then
     ui_print " "; ui_print "Repacking ramdisk failed. Aborting..."; exit 1;
   fi;
+  cd /tmp/anykernel;
   if [ -f "$bin/mkmtkhdr" ]; then
-    cd /tmp/anykernel;
     $bin/mkmtkhdr --rootfs ramdisk-new.cpio.gz;
     mv -f ramdisk-new.cpio.gz-mtk ramdisk-new.cpio.gz;
     case $kernel in
@@ -126,17 +133,27 @@ write_boot() {
       *) $bin/mkmtkhdr --kernel $kernel; kernel=$kernel-mtk;;
     esac;
   fi;
-  $bin/mkbootimg --kernel $kernel --ramdisk /tmp/anykernel/ramdisk-new.cpio.gz $second --cmdline "$cmdline" --board "$board" --base $base --pagesize $pagesize --kernel_offset $kerneloff --ramdisk_offset $ramdiskoff $secondoff --tags_offset "$tagsoff" --os_version "$osver" --os_patch_level "$oslvl" $dtb --output /tmp/anykernel/boot-new.img;
+  $bin/mkbootimg --kernel $kernel --ramdisk ramdisk-new.cpio.gz $second --cmdline "$cmdline" --board "$board" --base $base --pagesize $pagesize --kernel_offset $kerneloff --ramdisk_offset $ramdiskoff $secondoff --tags_offset "$tagsoff" --os_version "$osver" --os_patch_level "$oslvl" $dtb --output boot-new.img;
   if [ $? != 0 ]; then
     ui_print " "; ui_print "Repacking image failed. Aborting..."; exit 1;
-  elif [ `wc -c < /tmp/anykernel/boot-new.img` -gt `wc -c < /tmp/anykernel/boot.img` ]; then
+  elif [ `wc -c < boot-new.img` -gt `wc -c < boot.img` ]; then
     ui_print " "; ui_print "New image larger than boot partition. Aborting..."; exit 1;
+  fi;
+  if [ -f "$bin/futility" -a -d "$bin/chromeos" ]; then
+    $bin/futility vbutil_kernel --pack boot-new-signed.img --keyblock $bin/chromeos/kernel.keyblock --signprivate $bin/chromeos/kernel_data_key.vbprivk --version 1 --vmlinuz boot-new.img --bootloader $bin/chromeos/empty --config $bin/chromeos/empty --arch arm --flags 0x1;
+    if [ $? != 0 ]; then
+      ui_print " "; ui_print "Signing image failed. Aborting..."; exit 1;
+    fi;
+    mv -f boot-new-signed.img boot-new.img;
   fi;
   if [ -f "/data/custom_boot_image_patch.sh" ]; then
     ash /data/custom_boot_image_patch.sh /tmp/anykernel/boot-new.img;
     if [ $? != 0 ]; then
       ui_print " "; ui_print "User script execution failed. Aborting..."; exit 1;
     fi;
+  fi;
+  if [ "$(strings /tmp/anykernel/boot.img | grep SEANDROIDENFORCE )" ]; then
+    printf 'SEANDROIDENFORCE' >> /tmp/anykernel/boot-new.img;
   fi;
   if [ -f "$bin/flash_erase" -a -f "$bin/nandwrite" ]; then
     $bin/flash_erase $block 0 0;
@@ -260,6 +277,18 @@ patch_fstab() {
     esac;
     newentry=$(echo "$entry" | sed "s;${part};${6};");
     sed -i "s;${entry};${newentry};" $1;
+  fi;
+}
+
+# patch_cmdline <cmdline match string> [<replacement string>]
+patch_cmdline() {
+  cmdfile=`ls $split_img/*-cmdline`;
+  if [ -z "$(grep "$1" $cmdfile)" ]; then
+    cmdtmp=`cat $cmdfile`;
+    echo "$cmdtmp $1" > $cmdfile;
+  else
+    match=$(grep -o "$1.*$" $cmdfile | cut -d\  -f1);
+    sed -i -e "s;${match};${2};" -e 's;  ; ;' -e 's;[ \t]*$;;' $cmdfile;
   fi;
 }
